backend/pyproject.toml
[project]
name = "soip-v2"
version = "0.1.0"
description = "Student Opportunity Intelligence Platform"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "beautifulsoup4>=4.12.0",
    "crawl4ai>=0.4.0",
    "markdownify>=0.14.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
